advice_bot/requirements.txt
python-telegram-bot==21.4
requests==2.32.3
python-dotenv==1.0.1
openai>=1.51.0